requirements.txt
beautifulsoup4==4.11.1
dominate==2.6.0
numpy==1.21.2
opencv_python==4.5.5.64
Pillow==9.1.0
requests==2.27.1
torch==1.11.0
torchvision==0.12.0
visdom==0.1.8.9
